Fix taxonomy term creation
`parent` may be empty if multiple taxonomies are created in a single
`Given` block.

Additionally `id` is not the right base field for the query but it
should be `name`.

Finally the entityQuery does not return loaded entities but returns IDs
so calling `->id()` on it will break as well.